docs(tutorial/2 - Angular Templates): describe your change...
In step 2

In Controller.js, the name for var phonecatApp is same as the module.
So it sometimes gives following error ::

Uncaught Error: [$injector:modulerr] Failed to instantiate module phonecatApp due to:
Error: [$injector:nomod] Module 'phonecatApp' is not available! You either misspelled the module name or forgot to load it. If registering a module ensure that you specify the dependencies as the second argument.

This can be handled by giving different name for var and module.


Regards,
Shifali